[renpy-build-action] Add check for empty old-game directory
In newer versions of Ren'Py, the old-game directory was introduced.
Per a suggestion (#5), this commit adds a check to verify that, if
the old-game dir exists, that it isn't empty. If it is, the script
exits with return code 1, which should trigger GitHub Actions to
treat the workflow as a failed run.

Signed-off-by: Marquis Kurt <software@marquiskurt.net>